Add sass script and theme CSS
Add a "sass" script to the package.json to compile the style.scss file.
It's not perfect and won't do it on editing but it should be good enough
for my purposes (one file).

Copy the CSS from the old jekyll theme into a new theme.css assest and
include the stylesheet in the default template.